+++ app/db/base.py
@@ -1,0 +1,5 @@
+# Import all the models, so that Base has them before being
+# imported by Alembic
+from app.db.base_class import Base  # noqa
+from app.models.job import Job  # noqa
+from app.models.user import User  # noqa