+++ app/models/__init__.py
@@ -1,0 +1,2 @@
+from .job import Job  # noqa
+from .user import User  # noqa