+++ app/routers/views.py
@@ -1,17 +1,26 @@
-import flux.job
-from fastapi import APIRouter, Depends, Request
-from fastapi.responses import HTMLResponse, RedirectResponse
-from fastapi.templating import Jinja2Templates
+from datetime import timedelta
 
+import flux.job
+from fastapi import APIRouter, Depends, HTTPException, Request, Response
+from fastapi.responses import HTMLResponse, RedirectResponse
+from fastapi.security import OAuth2PasswordRequestForm
+from fastapi.templating import Jinja2Templates
+from sqlalchemy.orm import Session
+
+import app.core.security as security
 import app.library.flux as flux_cli
 import app.library.helpers as helpers
 import app.library.launcher as launcher
+import app.routers.depends as deps
 from app.core.config import settings
+from app.crud import user as crud_user
 from app.forms import SubmitForm
 from app.library.auth import check_auth
 
 # These views never have auth!
 router = APIRouter(tags=["views"])
+
+templates = Jinja2Templates(directory="templates/")
 
 # These views do :)
 auth_views_router = APIRouter(
@@ -23,6 +32,27 @@ templates = Jinja2Templates(directory="templates/")
 
 # Require auth (and the user in the view)
 user_auth = Depends(check_auth) if settings.require_auth else None
+
+
+@router.post(f"/{deps.login_url}")
+async def login(
+    form_data: OAuth2PasswordRequestForm = Depends(), db: Session = Depends(deps.get_db)
+):
+    """
+    This is the API endpoint to request an authentication token.
+    """
+    user = crud_user.authenticate(
+        db, user_name=form_data.username, password=form_data.password
+    )
+    if not user:
+        raise HTTPException(status_code=400, detail="Incorrect email or password")
+    elif not crud_user.is_active(user):
+        raise HTTPException(status_code=400, detail="Inactive user")
+    access_token_expires = timedelta(minutes=settings.access_token_expires_minutes)
+    access_token = security.create_access_token(
+        user.id, expires_delta=access_token_expires
+    )
+    return {"access_token": access_token, "token_type": "Bearer"}
 
 
 @router.get("/", response_class=HTMLResponse)
@@ -49,6 +79,27 @@ async def jobs_table(request: Request, user=user_auth):
         {
             "request": request,
             "jobs": jobs,
+        },
+    )
+
+
+@router.get("/logout")
+async def logout(request: Request, response: Response):
+    """
+    This isn't entirely working yet.
+
+    I usually open a new tab/window to reset basic auth. We likely
+    need a logout button to be handled somehow in javascript.
+    """
+    response.delete_cookie("basic")
+    response.delete_cookie("bearer")
+    response.delete_cookie("access_token")
+    data = helpers.get_page("index.md")
+    return templates.TemplateResponse(
+        "index.html",
+        {
+            "request": request,
+            "data": data,
         },
     )
 
@@ -87,7 +138,8 @@ async def job_info(request: Request, jobid, msg=None, user=user_auth):
 
 # Submit a job via a form
 @auth_views_router.get("/jobs/submit", response_class=HTMLResponse)
-async def submit_job(request: Request, _=user_auth):
+async def submit_job(request: Request, user=user_auth):
+    print(user)
     form = SubmitForm(request)
     return templates.TemplateResponse(
         "jobs/submit.html",
@@ -121,7 +173,9 @@ async def submit_job_post(request: Request, user=user_auth):
         print(form.kwargs)
 
         if form.kwargs.get("is_launcher") is True:
-            messages.append(launcher.launch(form.kwargs, workdir=form.workdir))
+            messages.append(
+                launcher.launch(form.kwargs, workdir=form.workdir, user=user)
+            )
         else:
             return submit_job_helper(request, app, form, user=user)
     else:
@@ -143,10 +197,10 @@ def submit_job_helper(request, app, form, user):
     A helper to submit a flux job (not a launcher)
     """
     # Submit the job and return the ID, but allow for error
-    if 1 == 1:
-        # Prepare the flux job! We don't support envars here yet
+    # Prepare the flux job! We don't support envars here yet
+    try:
         fluxjob = flux_cli.prepare_job(
-            form.kwargs, runtime=form.runtime, workdir=form.workdir
+            user, form.kwargs, runtime=form.runtime, workdir=form.workdir
         )
         flux_future = flux_cli.submit_job(app.handle, fluxjob, user=user)
         jobid = flux_future.get_id()
@@ -160,8 +214,8 @@ def submit_job_helper(request, app, form, user):
                 "messages": [message],
             },
         )
-    # except Exception as e:
-    #    form.errors.append("There was an issue submitting that job: %s" % str(e))
+    except Exception as e:
+        form.errors.append("There was an issue submitting that job: %s" % str(e))
 
     return templates.TemplateResponse(
         "jobs/submit.html",