+++ tests/test_api.py
@@ -3,13 +3,14 @@ import os
 import sys
 import time
 
-import pytest
 from fastapi.testclient import TestClient
 
 here = os.path.abspath(os.path.dirname(__file__))
 root = os.path.dirname(here)
 
 sys.path.insert(0, root)
+
+from jose import jwt  # noqa
 
 from app.main import app  # noqa
 
@@ -27,62 +28,91 @@ def get_basic_auth(username, password):
 
 
 def get_headers():
+    """
+    Prepare JWT encoded header.
+
+    This is encoded with the shared user/server secret
+    """
     flux_user = os.environ.get("FLUX_USER")
     flux_token = os.environ.get("FLUX_TOKEN")
-    auth_header = get_basic_auth(flux_user, flux_token)
-    if isinstance(auth_header, bytes):
-        auth_header = auth_header.decode("utf-8")
-    return {"Authorization": "Basic %s" % auth_header}
+    secret_key = os.environ.get("FLUX_SECRET_KEY")
+    print(flux_user)
+    print(flux_token)
+    print(secret_key)
+    if not secret_key:
+        sys.exit("Cannot generate header without secret key")
+    auth_str = {"user": flux_user, "pass": flux_token, "scope": "token"}
+    encoded = jwt.encode(auth_str, secret_key)  # algorithm is 'HS256'
+    return {"Authorization": f"Bearer {encoded}"}
 
 
 # Do we want auth?
+access_token = None
 test_auth = False
-test_pam_auth = False
-test_pam_auth_fail = False
 headers = {}
 
 if os.environ.get("TEST_AUTH"):
     test_auth = True
-    headers = get_headers()
 
-if os.environ.get("TEST_PAM_AUTH"):
-    test_pam_auth = True
-    test_pam_auth_fail = os.environ["TEST_PAM_AUTH_FAIL"] == "true"
-    headers = get_headers()
-
-skip_if_expected_fail = pytest.mark.skipif(test_pam_auth_fail, reason="Unix stuff")
+methods = {"get": client.get, "post": client.post}
 
 
-@skip_if_expected_fail
+def authenticate(endpoint, method="get", json=None, params=None, expected_status=200):
+    global access_token
+
+    response = methods[method](endpoint, json=json, params=params)
+    if test_auth and not access_token and response.status_code != 401:
+        raise ValueError(f"{endpoint} with auth enabled should return 401")
+
+    # We should only have to get once and reuse
+    if test_auth and not access_token:
+        if "www-authenticate" not in response.headers:
+            raise ValueError(f"{endpoint} missing www-authenticate header.")
+
+        response = client.post("/v1/token", headers=get_headers())
+        assert response.status_code == 200
+        token = response.json()
+        assert "access_token" in token and "token_type" in token
+        access_token = token["access_token"]
+
+    if test_auth:
+        # Redo the request with the token
+        response = methods[method](
+            endpoint,
+            headers={"Authorization": f"Bearer {access_token}"},
+            json=json,
+            params=params,
+        )
+    assert response.status_code == expected_status
+    return response
+
+
 def test_submit_list_job():
     """
     Test a manual submission
     """
     # No jobs are submit!
-    response = client.get("/v1/jobs", headers=headers)
-    assert response.status_code == 200
+    response = authenticate("/v1/jobs")
+
     assert not response.json()["jobs"]
 
     # Now submit a job, ensure we get one job back
-    response = client.post(
-        "/v1/jobs/submit", json={"command": "sleep 5"}, headers=headers
+    response = authenticate(
+        "/v1/jobs/submit", method="post", json={"command": "sleep 5"}
     )
-    assert response.status_code == 200
     result = response.json()
     assert "id" in result
     jobid = result["id"]
 
     # List jobs with and without details
-    response = client.get("/v1/jobs", headers=headers)
-    assert response.status_code == 200
+    response = authenticate("/v1/jobs")
     result = response.json()
     assert "jobs" in result
     result = result["jobs"]
     assert len(result) == 1
     assert result[0]["id"] == jobid
 
-    response = client.get("/v1/jobs", headers=headers, json={"details": True})
-    assert response.status_code == 200
+    response = authenticate("/v1/jobs", json={"details": True})
     result = response.json()
     assert str(jobid) in result
     result = result[str(jobid)]
@@ -92,92 +122,77 @@ def test_submit_list_job():
     assert result["id"] == jobid
 
     # Testing asking for a listing
-    response = client.get(
-        "/v1/jobs", headers=headers, json={"details": True, "listing": True}
-    )
-    assert response.status_code == 200
+    response = authenticate("/v1/jobs", json={"details": True, "listing": True})
     result = response.json()
     assert isinstance(result, list)
     assert result[0]["id"] == jobid
 
     # And setting an integer limit (only relevant for details)
-    response = client.get(
-        "/v1/jobs", headers=headers, json={"limit": 0, "details": True}
-    )
-    assert response.status_code == 200
+    response = authenticate("/v1/jobs", json={"limit": 0, "details": True})
     result = response.json()
     assert not result
 
 
-@skip_if_expected_fail
 def test_cancel_job():
     # Now submit a job, ensure we get one job back
-    response = client.post(
-        "/v1/jobs/submit", json={"command": "sleep 10"}, headers=headers
+    response = authenticate(
+        "/v1/jobs/submit", method="post", json={"command": "sleep 10"}
     )
-    assert response.status_code == 200
     result = response.json()
     jobid = result["id"]
-    response = client.post(f"/v1/jobs/{jobid}/cancel", headers=headers)
-    assert response.status_code == 200
+    response = authenticate(f"/v1/jobs/{jobid}/cancel", method="post")
     # TODO we don't have way to actually verify that cancel happened
 
 
-@skip_if_expected_fail
 def test_submit_option_flags():
     """
     Test that option flags are parsed.
     """
     # Submit a job with invalid flags
-    response = client.post(
+    response = authenticate(
         "/v1/jobs/submit",
+        method="post",
         json={"command": "sleep 1", "option_flags": "invalid"},
-        headers=headers,
+        expected_status=400,
     )
-    assert response.status_code == 400
     result = response.json()
     errors = result.get("Errors")
     assert errors
     assert "is invalid" in errors[0]
 
     # Another invalid pattern
-    response = client.post(
+    response = authenticate(
         "/v1/jobs/submit",
+        method="post",
         json={"command": "sleep 1", "option_flags": {"-ompi": "nope"}},
-        headers=headers,
+        expected_status=400,
     )
-    assert response.status_code == 400
     result = response.json()
     errors = result.get("Errors")
     assert errors
     assert "keys without -o" in errors[0]
 
     # Valid
-    response = client.post(
+    response = authenticate(
         "/v1/jobs/submit",
+        method="post",
         json={"command": "sleep 1", "option_flags": {"ompi": "openmpi@5"}},
-        headers=headers,
     )
-    assert response.status_code == 200
     result = response.json()
     assert "id" in result
 
 
-@skip_if_expected_fail
 def test_job_output():
     """
     Test endpoint to retrieve list of job output
     """
-
     # Now submit a job, ensure we get one job back
-    response = client.post(
-        "/v1/jobs/submit", json={"command": "echo pancakes 🥞️🥞️🥞️"}, headers=headers
+    response = authenticate(
+        "/v1/jobs/submit", method="post", json={"command": "echo pancakes 🥞️🥞️🥞️"}
     )
-    assert response.status_code == 200
     result = response.json()
     jobid = result["id"]
-    res = client.get(f"/v1/jobs/{jobid}/output", headers=headers)
-    assert response.status_code == 200
+    res = authenticate(f"/v1/jobs/{jobid}/output")
     lines = res.json()
 
     # First try, often we won't have output yet
@@ -186,24 +201,20 @@ def test_job_output():
     time.sleep(3)
 
     # Try again - we should have it after a sleep
-    res = client.get(f"/v1/jobs/{jobid}/output", headers=headers)
-    assert response.status_code == 200
+    res = authenticate(f"/v1/jobs/{jobid}/output")
     lines = res.json()
     assert "Output" in lines
     assert "pancakes 🥞️🥞️🥞️\n" in lines["Output"]
 
 
-@skip_if_expected_fail
 def test_job_query():
     """
     Test endpoint to query jobs
     """
     # Submit 5 jobs
     for _ in range(5):
-        client.post("/v1/jobs/submit", json={"command": "sleep 1"}, headers=headers)
-
-    response = client.get("/v1/jobs/search", headers=headers)
-    assert response.status_code == 200
+        authenticate("/v1/jobs/submit", method="post", json={"command": "sleep 1"})
+    response = authenticate("/v1/jobs/search")
     result = response.json()
 
     for key in ["recordsTotal", "recordsFiltered", "draw", "data"]:
@@ -212,15 +223,13 @@ def test_job_query():
     assert len(result["data"]) == total
 
     # Ask to start at 2 (should be one less record)
-    response = client.get("/v1/jobs/search", params={"start": 1}, headers=headers)
-    assert response.status_code == 200
+    response = authenticate("/v1/jobs/search", params={"start": 1})
     result = response.json()
     assert result["recordsFiltered"] == total - 1
     assert result["recordsTotal"] == total
 
     # Ask for specific length
-    response = client.get("/v1/jobs/search", params={"length": 3}, headers=headers)
-    assert response.status_code == 200
+    response = authenticate("/v1/jobs/search", params={"length": 3})
     result = response.json()
     assert result["recordsFiltered"] == 3
     assert result["recordsTotal"] == total