+++ clients/python/CHANGELOG.md
@@ -14,6 +14,7 @@ and **Merged pull requests**. Critical items to know are:
 The versions coincide with releases on pip. Only major versions will be released as tags on Github.
 
 ## [0.0.x](https://github.com/flux-framework/flux-restful-api/tree/main) (0.0.x)
+ - refactor to require secret key, oauth2 flow (0.1.0)
  - add simple retry to requests (0.0.16)
  - support for adding option flags to submit (0.0.15)
  - support for `is_launcher` parameter to indicate a launcher should be used instead (0.0.14)